requirements.txt
numpy
pandas
statsmodels
matplotlib
ipywidgets
scikit-sklearn
